chore: rename kebab-case service files to PascalCase
- apicurio-registry.service.ts → ApicurioRegistryService.ts
- openapi-import.service.ts → OpenApiImportService.ts
- Move and rename corresponding test files into src/test/services/
- Update all imports across ImportOpenApiCommand.ts, OpenApiProvider.ts,
  and both test files
