+++ src/commands/ImportOpenApiCommand.ts
@@ -20,7 +20,7 @@ import {
 	KAOTO_OPENAPI_FILES_REGEXP_SETTING_ID,
 	KAOTO_REST_APICURIO_REGISTRY_URL_SETTING_ID,
 } from '../constants';
-import { ApicurioRegistryService, ApicurioRegistryUrlError, type ApicurioArtifact } from '../services/apicurio-registry.service';
+import { ApicurioRegistryService, ApicurioRegistryUrlError, type ApicurioArtifact } from '../services/ApicurioRegistryService';
 import { OpenApiImportService, OpenApiParseError, OpenApiValidationError, type ParsedOperation } from '../services/openapi-import.service';
 import { AbstractNewCamelRouteCommand } from './AbstractNewCamelRouteCommand';
